inital Birdcat implementation
Clones the bearcat so we can do changes to it as needed. Also adds bits of the Torch's overmap to it - Namely it's mining z level and spooky away mission, mostly for proper mining purposes. A proper mining array for the Birdcat will come soon.

Birdcat is now the inital map too.

Also moved around the bird skeleton.